python-worker/requirements.txt
psycopg2-binary==2.9.9
boto3==1.34.0
numpy==1.26.2
scipy==1.11.4
python-pptx==0.6.23
reportlab==4.0.7
weasyprint==60.1
python-dotenv==1.0.0
openpyxl==3.1.2
pandas==2.1.4
requests==2.31.0
